api/requirements.txt
tensorflow==2.17.0
keras==3.6.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.17.0
matplotlib
numpy
